[src] Rename file of MUSDB18 dataset
